# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface RunRequest {
  language: string;
  code: string;
  filename?: string;
}

export interface RunResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  duration: number;
  error?: string | null;
  html?: string | null;
}

export type ProjectBuildRequestType =
  (typeof ProjectBuildRequestType)[keyof typeof ProjectBuildRequestType];

export const ProjectBuildRequestType = {
  flutter: "flutter",
  "react-native": "react-native",
  android: "android",
  ios: "ios",
} as const;

export type ProjectBuildRequestFiles = { [key: string]: string };

export interface ProjectBuildRequest {
  type: ProjectBuildRequestType;
  files: ProjectBuildRequestFiles;
  name?: string;
}

export type ProjectBuildResponseStatus =
  (typeof ProjectBuildResponseStatus)[keyof typeof ProjectBuildResponseStatus];

export const ProjectBuildResponseStatus = {
  queued: "queued",
  building: "building",
  success: "success",
  failed: "failed",
  preview_ready: "preview_ready",
} as const;

export interface ProjectBuildResponse {
  jobId: string;
  status: ProjectBuildResponseStatus;
  previewUrl?: string | null;
  embedUrl?: string | null;
  qrUrl?: string | null;
  message?: string | null;
}

export type BuildResponseStatus =
  (typeof BuildResponseStatus)[keyof typeof BuildResponseStatus];

export const BuildResponseStatus = {
  queued: "queued",
  building: "building",
  success: "success",
  failed: "failed",
} as const;

export interface BuildResponse {
  jobId: string;
  status: BuildResponseStatus;
  queuePosition?: number;
}

export type JobStatusStatus =
  (typeof JobStatusStatus)[keyof typeof JobStatusStatus];

export const JobStatusStatus = {
  queued: "queued",
  building: "building",
  success: "success",
  failed: "failed",
} as const;

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  logs?: string | null;
  download?: string | null;
  queuePosition?: number | null;
  startedAt?: string | null;
  completedAt?: string | null;
  stage?: string | null;
  previewUrl?: string | null;
  embedUrl?: string | null;
  qrUrl?: string | null;
}

export interface BuildLogs {
  jobId: string;
  logs: string;
  stage?: string | null;
}

export type CreateProjectRequestFiles = { [key: string]: string };

export interface CreateProjectRequest {
  name: string;
  projectType?: string;
  files: CreateProjectRequestFiles;
}

export type UpdateProjectRequestFiles = { [key: string]: string };

export interface UpdateProjectRequest {
  name?: string;
  projectType?: string;
  files?: UpdateProjectRequestFiles;
}

export type ProjectFiles = { [key: string]: string } | null;

export interface Project {
  id: string;
  name: string;
  projectType: string;
  files?: ProjectFiles;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectSummary {
  id: string;
  name: string;
  projectType: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectListResponse {
  projects: ProjectSummary[];
}

export interface ProjectResponse {
  project: Project;
}

export interface ErrorResponse {
  error: string;
}

export type StartBuildBody = {
  project: Blob;
};

export type DeleteProject200 = {
  success: boolean;
  id: string;
};