# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /build:
    post:
      operationId: startBuild
      tags: [builds]
      summary: Start a Flutter APK build (ZIP upload)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                project:
                  type: string
                  format: binary
              required:
                - project
      responses:
        "200":
          description: Build queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /build/project:
    post:
      operationId: buildProject
      tags: [builds]
      summary: Build or preview from file map (react-native, flutter, android)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectBuildRequest"
      responses:
        "200":
          description: Build started or preview ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectBuildResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /status/{jobId}:
    get:
      operationId: getBuildStatus
      tags: [builds]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /download/{jobId}:
    get:
      operationId: downloadApk
      tags: [builds]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: APK binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /logs/{jobId}:
    get:
      operationId: getBuildLogs
      tags: [builds]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Build logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildLogs"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ProjectBuildRequest:
      type: object
      properties:
        type:
          type: string
          enum: [flutter, react-native, android, ios]
        files:
          type: object
          additionalProperties:
            type: string
        name:
          type: string
      required: [type, files]
    ProjectBuildResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, building, success, failed, preview_ready]
        previewUrl:
          type: string
          nullable: true
        embedUrl:
          type: string
          nullable: true
        qrUrl:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required: [jobId, status]
    BuildResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, building, success, failed]
        queuePosition:
          type: integer
      required: [jobId, status]
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, building, success, failed]
        logs:
          type: string
          nullable: true
        download:
          type: string
          nullable: true
        queuePosition:
          type: integer
          nullable: true
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        stage:
          type: string
          nullable: true
        previewUrl:
          type: string
          nullable: true
        embedUrl:
          type: string
          nullable: true
        qrUrl:
          type: string
          nullable: true
      required: [jobId, status]
    BuildLogs:
      type: object
      properties:
        jobId:
          type: string
        logs:
          type: string
        stage:
          type: string
          nullable: true
      required: [jobId, logs]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]