# lib/api-zod/src/generated/types/buildResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { BuildResponseStatus } from "./buildResponseStatus";

export interface BuildResponse {
  jobId: string;
  status: BuildResponseStatus;
  queuePosition?: number;
}