# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { JobStatusStatus } from "./jobStatusStatus";

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  logs?: string | null;
  download?: string | null;
  queuePosition?: number | null;
  startedAt?: string | null;
  completedAt?: string | null;
  stage?: string | null;
  previewUrl?: string | null;
  embedUrl?: string | null;
  qrUrl?: string | null;
}