# lib/api-zod/src/generated/types/projectBuildRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ProjectBuildRequestFiles } from "./projectBuildRequestFiles";
import type { ProjectBuildRequestType } from "./projectBuildRequestType";

export interface ProjectBuildRequest {
  type: ProjectBuildRequestType;
  files: ProjectBuildRequestFiles;
  name?: string;
}