# lib/api-zod/src/generated/types/projectBuildResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ProjectBuildResponseStatus } from "./projectBuildResponseStatus";

export interface ProjectBuildResponse {
  jobId: string;
  status: ProjectBuildResponseStatus;
  previewUrl?: string | null;
  embedUrl?: string | null;
  qrUrl?: string | null;
  message?: string | null;
}